split test files and unuse dupl
